backend/pyproject.toml
[tool.poetry]
name = "online-shop"
version = "0.1.0"
description = ""
authors = ["poslam <kirden04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-phonenumber-field = "^7.2.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
